% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{crossDist}
\alias{crossDist}
\title{Computed the Euclidian Distance Matrix}
\usage{
crossDist(coord1, coord2 = coord1)
}
\arguments{
\item{coord1, coord2}{Matrices with the coordinates of locations, between
which distances are to be computed.}
}
\value{
A \code{dim(coord1)[1]}-by-\code{dim(coord2)[1]} distance matrix.
}
\description{
Computed the Euclidian distance matrix between to sets of points.
}
\author{
Johan Lindström
}
\concept{basic linear algebra}
\concept{covariance functions}
