% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatTree.R
\name{spatTree}
\alias{spatTree}
\title{spatTree}
\usage{
spatTree(
  Y,
  X,
  domain = rep(1, length(Y)),
  X.fix = NULL,
  coords = NULL,
  cov.type = NULL,
  cov.opts = NULL,
  i.Sig = NULL,
  m = ceiling(dim(X)[2]/3),
  r = 5,
  max.depth = NULL,
  lambda = 0,
  pen.fix = TRUE,
  L = matrix(0, length(unique(domain)), length(unique(domain)))
)
}
\arguments{
\item{Y}{Observed values}

\item{X}{Covariates used to construct the spatially adjusted regression tree}

\item{domain}{Used to construct multiple regression trees simultaneously. For
details, see Hee Wai et al., 2020}

\item{X.fix}{Covariates to fit as linear fixed effects in addition to the
spatially regression tree (Optional)}

\item{coords}{x and y coordinates for locations of the points}

\item{cov.type}{Type of covariance function to use.  Currently supported are
'exp' (Exponential), 'LRK' (low rank kriging), and 'TPRS',
thin plate regression spline basis functions}

\item{cov.opts}{Parameters for the covariance function.  See additional
documentation for more details}

\item{i.Sig}{Alternative to specifying coords, cov.type, and cov.opts allows
specification of an inverse covariance matrix.  If all are not
specified assumes independence}

\item{m}{Number of covariates to randomly sample for each split of
the spatially adjusted regression tree. Applicable only for use in
Random Spatial Forests}

\item{r}{Minimum Leaf Size}

\item{max.depth}{Maximum number of splits to construct (if applicable)}

\item{lambda}{Penalty parameter for L. Applicable only is multiple domains
are used.}

\item{pen.fix}{logical indicating whether or not to penalize the fixed
effects in addition to the tree estimates.  Applicable only if
multiple domains are used.}

\item{L}{Penalty matrix encouraging similarity among estimates in different
domains.  Applicable only is multiple domains are used.}
}
\value{
a list of result with ...
}
\description{
spatTree
}
