% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatRF.R
\name{spatRF}
\alias{spatRF}
\title{Title}
\usage{
spatRF(
  pars,
  Y,
  X,
  domain,
  coords,
  group = 1:length(Y),
  X.fix = NULL,
  cov.type = "exp",
  cov.opts = NULL,
  X.test = NULL,
  X.fix.test = NULL,
  test.coords = NULL,
  test.domain = NULL,
  t = 500,
  m = floor(dim(X)[2]/3),
  r = 5,
  k = floor(length(Y)/r),
  replace = TRUE,
  s = ifelse(replace, length(Y), ceiling(0.632 * length(Y))),
  var.imp = FALSE,
  imp.msr = c("perm", "imp"),
  pen.fix = TRUE,
  L = matrix(0, length(unique(domain)), length(unique(domain))),
  scale = FALSE,
  ...
)
}
\arguments{
\item{pars}{}

\item{Y}{}

\item{X}{}

\item{domain}{}

\item{coords}{}

\item{group}{}

\item{X.fix}{}

\item{cov.type}{}

\item{cov.opts}{}

\item{X.test}{}

\item{X.fix.test}{}

\item{test.coords}{}

\item{test.domain}{}

\item{t}{}

\item{m}{}

\item{r}{}

\item{k}{}

\item{replace}{}

\item{s}{}

\item{var.imp}{}

\item{imp.msr}{}

\item{pen.fix}{}

\item{L}{}

\item{scale}{}

\item{...}{}
}
\value{

}
\description{
Title
}
